+++ src/DomOptions.php
@@ -15,27 +15,38 @@ class DomOptions
     /** @var bool */
     private $omitRootElement;
 
+    /** @var bool */
+    private $useAttributeNodeName;
+
     public const SKIP_ATTRIBUTES = 'skip_attributes';
 
     public const KEEP_MIXED_CONTENT = 'keep_mixed_content';
 
     public const OMIT_ROOT_ELEMENT = 'omit_root_element';
 
-    private function __construct(bool $skipAttributes, bool $keepMixedContent, bool $omitRootElement)
-    {
+    public const USE_ATTRIBUTE_NODE_NAME = 'use_attribute_node_name';
+
+    private function __construct(
+        bool $skipAttributes,
+        bool $keepMixedContent,
+        bool $omitRootElement,
+        bool $useAttributeNodeName
+    ) {
         $this->skipAttributes = $skipAttributes;
         $this->keepMixedContent = $keepMixedContent;
         $this->omitRootElement = $omitRootElement;
+        $this->useAttributeNodeName = $useAttributeNodeName;
     }
 
-    /** @param array{skip_attributes?: bool, keep_mixed_content?: bool, omit_root_element?: bool} $array */
+    /** @param array{skip_attributes?: bool, keep_mixed_content?: bool, omit_root_element?: bool, use_attribute_node_name?: bool} $array */
     public static function fromArray(array $array): self
     {
         $skipAttributes = (bool) ($array[self::SKIP_ATTRIBUTES] ?? false);
         $keepMixedContent = (bool) ($array[self::KEEP_MIXED_CONTENT] ?? false);
         $omitRootElement = (bool) ($array[self::OMIT_ROOT_ELEMENT] ?? false);
+        $useAttributeNodeName = (bool) ($array[self::USE_ATTRIBUTE_NODE_NAME] ?? false);
 
-        return new self($skipAttributes, $keepMixedContent, $omitRootElement);
+        return new self($skipAttributes, $keepMixedContent, $omitRootElement, $useAttributeNodeName);
     }
 
     public function getSkipAttributes(): bool
@@ -51,5 +62,10 @@ class DomOptions
     public function getOmitRootElement(): bool
     {
         return $this->omitRootElement;
+    }
+
+    public function getUseAttributeNodeName(): bool
+    {
+        return $this->useAttributeNodeName;
     }
 }